Buoyancy submodule
